Added AI and it's relevant classes
Added all the AI Scripts, now with updated class modularization. Still need to extract the Flee function but this will be coming soon
Changed the Namespace to fit in with the new theme. All of my AI work now falls under the `TrojanMouse.AI` namespace, with movement and behaviours being subcatagories