Added basic costs, cleaned up navmesh
Added basic water nav costs to the water, it's currently at 5 so the AI favours going over the bridges. Did run into a bug with the terrain where the it's too steep for the AI to climb or path which stops them from working entirely as they can't reach the home region.

Cleaned up the navmesh a little bit, readded the ignore from build to the rocks to make it more performant, going to keep trying to clean it up