More work on drawbridge stuff
updated DrawBridge.cs, movement of the bridges is now in, next is to get the navmesh to change
changed prefab bridge to have the components
made changes to the scene, rebaked the navmesh - gruttels can no longer walk on water, for this fix needed to introduce changes to the big water plane. There's now a near one which included in the bake and a far one that's not included. Think like a skybox.