Progress on machine filling
Machine should now increase fill amount when rubbish is deposited, when full it will call the reward function of the reward set to it. Those functions are just debug statements right now but they work. Also created a  shader and temporary dome object that fill as rubbish amount increases, currently not working.